Expanded colors validation and parsing logic
Now we check each color to see if it's in our tuple of ANSI_COLOR_NAMES.
If not, it must be hex (we're not dealing with RGB for now).

- Added new function 'validate_hex' to validate a hex. Only 6-letter
hexes are accepted. May add CSS shorthand in the future.
- Added new function 'parse_colors' to manage determination of ANSI vs.
hex, and modify the input string accordingly (mainly to make inputs
flexibly accept hex with or without a preceding '#')