Re-added 'colors' interface
This is not an ideal solution, but it's the one we'll go with for now in
the interest of making our code modular. Previously, the logic for
settings colors was executed in the args module. I've re-added this
interface so we can properly test the way colors are set.

This solution involves changing the 'colors' attribute for the
ConwayState instance as well as the class attribute, so newly generated
instances (i.e., those returned from the change_state method) will
inherit the correct colors.

Another way to manage this might be to override __new__, but this would
make the underlying Protocol signature confusing.

More to come.