Improved performance by building next generation dynamically
Before, we were creating the 'next_generation' (previously referred to
as 'ref') by creating a copy of the existing matrix and overwriting
it. This is incredibly inefficient.

Replaced that logic with a generation that's created dynamically,
starting with and empty list of list of StateData and buliding on it
as we iterate through the matrix.