Modified convenience methods in __init__ to meet new functionality
Without a Console object to drive default xmax/ymax in the absence of
user-provided values, we'll no longer utilize the 'spawn' method to
spawn on automaton inside of another equal to the size of the
console.

Until we have algorithms that allow for pseudo-infinite space, the
user will be responsible for defining the bounds of the simulation.